Switching to a safe version of NewChunk#bufX method due to crashes on Jenkins
Jenkins builds showed the default implementation of NewChunk#bufX
can sometimes cause segfaults. We have seen similar behavior in AutoML
testing (rarely).

This was reported as a JDK bug, however, due to use of Unsafe API
the bug was rejected. We are providing an implementation of the bugX
method that doesn't rely on Unsafe and is turned on by default in this change.

(cherry picked from commit 002d26e8cf3067445e913f6186cb9423c4aae2b1)