Fixed a test that expects exact value of MPCE (details in body)
The model is trained on a dataset that contains NA in response. In this
PR we fixed a related bug and we no longer learn from observations
that have NA label. It is thus expected that tests that have very specific
expectation will need to be adapted.

(cherry picked from commit 166bbd4a435fe5883288198bcac14a4e301cf4ce)